refactor: 패키지를 세분화하여 분리
