--- domain/src/main/kotlin/domain/CoordinateGenerator.kt
@@ -1,5 +1,0 @@
-package domain
-
-fun interface CoordinateGenerator {
-    fun read(): Coordinate
-}
