+++ domain/src/test/kotlin/domain/BoardTest.kt
@@ -1,5 +1,6 @@
 package domain
 
+import domain.omokrule.RenjuRule
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
