+++ domain/src/test/kotlin/domain/RenjuRuleTest.kt
@@ -1,5 +1,6 @@
 package domain
 
+import domain.omokrule.RenjuRule
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
